#210 [feat] junit5 useCase test example
