--- domain/src/main/java/hous/release/domain/entity/MainTodo.kt
@@ -1,7 +1,0 @@
-package hous.release.domain.entity
-
-data class MainTodo(
-    val isNew: Boolean,
-    override val id: Int,
-    override val name: String
-): Rule(id, name)
