--- domain/src/main/java/hous/release/domain/entity/Rule.kt
@@ -1,6 +1,0 @@
-package hous.release.domain.entity
-
-abstract class Rule(
-    open val id: Int,
-    open val name: String
-)
